WEBAPP-231: Move getLanguagePath
